backend/requirements.txt
pydantic-settings==2.5.2
aiogram==3.24.0
pydantic==2.9.2
pillow==12.1.0
httpx==0.28.1